some code is moved handler->agent
